requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
contourpy==1.2.0
cryptography==41.0.5
cycler==0.12.1
daphne==4.0.0
distlib==0.3.8
distro==1.9.0
Django==5.1.2
django-bootstrap5==23.3
django-cors-headers==4.3.1
django-filter==24.3
django-jquery==3.1.0
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
exceptiongroup==1.2.0
Faker==30.8.1
filelock==3.13.3
fonttools==4.49.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
kafka-python==2.0.2
kiwisolver==1.4.5
matplotlib==3.8.3
msgpack==1.0.7
numpy==1.26.4
openai==1.12.0
packaging==23.2
pika==1.3.2
Pillow==10.1.0
platformdirs==4.2.0
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.2
pydantic_core==2.16.3
PyJWT==2.8.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.2
razorpay==1.4.1
redis==5.0.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
service-identity==23.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
stripe==11.2.0
tqdm==4.66.2
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.7
uvicorn==0.24.0.post1
virtualenv==20.25.1
zope.interface==6.1
